src/requirements.txt
bidict==0.21.2
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dnspython==1.16.0
eventlet==0.26.1
firebase-admin==4.5.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-MQTT==1.1.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.5.1
gevent==21.1.2
gevent-websocket==0.10.1
geopy==2.1.0
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.0
google-cloud-storage==1.37.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.37.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
monotonic==1.6
msgpack==1.0.2
numpy==1.19.5
packaging==20.9
paho-mqtt==1.5.1
pandas==1.2.0
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
SQLAlchemy==1.4.11
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.26.4
uuid==1.30
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
